@@ -1,13 +1,12 @@
-package com.sheryians.major.controller;
+package com.ecommerce.miniproject.controller;
 
-import com.sheryians.major.service.CategoryService;
-import com.sheryians.major.service.ProductService;
+import com.ecommerce.miniproject.service.CategoryService;
+import com.ecommerce.miniproject.service.ProductService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
-
 @Controller
 
 public class HomeController {
@@ -15,8 +14,6 @@ public class HomeController {
     CategoryService categoryService;
     @Autowired
     ProductService productService;
-
-
     @GetMapping({"/","home"})
     public String home(Model model){
         return "index";
@@ -31,8 +28,12 @@ public class HomeController {
     public String shopByCategory(@PathVariable int id, Model model){
         model.addAttribute("categories",categoryService.getAllCategory());
         model.addAttribute("products",productService.getAllProductsByCategory_id(id));
-        System.out.println(categoryService.getAllCategory());
-
         return "shop";
     }
+    @GetMapping("/shop/viewproduct/{id}")
+    public String viewProduct(@PathVariable int id,Model model){
+        model.addAttribute("product",productService.getProductById(id).get());
+        return "viewProduct";
+    }
+
 }
