@@ -1,11 +1,7 @@
-package com.sheryians.major.dto;
+package com.ecommerce.miniproject.dto;
 
-import com.sheryians.major.model.Category;
 import lombok.Data;
 
-import javax.persistence.FetchType;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
 @Data
 public class ProductDTO {
     private Long id;
