@@ -1,4 +1,4 @@
-package com.sheryians.major;
+package com.ecommerce.miniproject;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
